==> dev/4_make_pseudobulk.R <==
library(tidyverse)
library(targets)
library(glue)

library(arrow)
library(dplyr)
library(duckdb)
# Get input

result_directory = "/vast/scratch/users/mangiola.s/pseudobulk_cellNexus_1_0_4"

tar_script({
  
  result_directory = "/vast/scratch/users/mangiola.s/pseudobulk_cellNexus_1_0_4"
  
  
  #-----------------------#
  # Input
  #-----------------------#
  library(tidyverse)
  library(targets)
  library(tarchetypes)
  library(glue)
  library(qs)
  library(crew)
  library(crew.cluster)
  
  #-----------------------#
  # Packages
  #-----------------------#
  tar_option_set(
    packages = c(
      "CuratedAtlasQueryR", "stringr", "tibble", "tidySingleCellExperiment", "dplyr", "Matrix",
      "Seurat", "glue", "qs",  "purrr", "tidybulk", "tidySummarizedExperiment", "edgeR", 
      "crew", "magrittr", "digest", "readr", "forcats", "scuttle", "BiocParallel", "SummarizedExperiment"
    ),
    
    memory = "transient", 
    garbage_collection = 100, 
    storage = "worker", 
    retrieval = "worker", 
    error = "continue", 
    cue = tar_cue(mode = "never"), 
    workspace_on_error = TRUE,
    format = "qs",
    
    #-----------------------#
    # SLURM
    #-----------------------#
    controller = crew_controller_group(
      
      
      crew_controller_slurm(
        name = "slurm_1_5",
        slurm_memory_gigabytes_per_cpu = 5,
        slurm_cpus_per_task = 1,
        workers = 400,
        tasks_max = 20,
        verbose = T, 
        seconds_idle = 30
      ),
      crew_controller_slurm(
        name = "slurm_1_10",
        slurm_memory_gigabytes_per_cpu = 10,
        slurm_cpus_per_task = 1,
        workers = 300,
        tasks_max = 20,
        verbose = T, 
        seconds_idle = 30
      ),
      crew_controller_slurm(
        name = "slurm_1_20",
        script_lines = "#SBATCH --mem 20G",
        slurm_cpus_per_task = 1,
        workers = 200,
        tasks_max = 20,
        verbose = T, 
        seconds_idle = 30
      ),
      crew_controller_slurm(
        name = "elastic",
        workers = 500,
        tasks_max = 20,
        seconds_idle = 30,
        crashes_error = 10,
        options_cluster = crew_options_slurm(
          memory_gigabytes_required = c(20, 40, 80, 160), 
          cpus_per_task = c(2, 5, 10, 20), 
          time_minutes = c(30, 30, 60*4, 60*24),
          verbose = T
        )
      ),
      crew_controller_slurm(
        name = "slurm_1_80_1_core",
        script_lines = "#SBATCH --mem 80G",
        slurm_cpus_per_task = 1,
        workers = 100,
        tasks_max = 20,
        verbose = T, 
        seconds_idle = 30
      ),
      crew_controller_slurm(
        name = "slurm_1_80",
        script_lines = "#SBATCH --mem 80G",
        slurm_cpus_per_task = 20,
        workers = 30,
        tasks_max = 1,
        verbose = T, 
        seconds_idle = 30
      ),
      crew_controller_slurm(
        name = "slurm_1_150",
        slurm_memory_gigabytes_per_cpu = 150,
        slurm_cpus_per_task = 1,
        workers = 2,
        tasks_max = 1,
        verbose = T, 
        seconds_idle = 30
      ),
      crew_controller_slurm(
        name = "slurm_2_400",
        slurm_memory_gigabytes_per_cpu = 400,
        slurm_cpus_per_task = 2,
        workers = 1,
        tasks_max = 1,
        verbose = T, 
        seconds_idle = 30
      )
    )
  )
  
  split_metadata = function(metadata_parquet){
    
    # CAQ_directory = "/vast/projects/cellxgene_curated/cellNexus"
    
    my_metadata =
      tbl(
        dbConnect(duckdb::duckdb(), dbdir = ":memory:"),
        sql(glue("SELECT * FROM read_parquet('{metadata_parquet}')"))
      ) |> 
      
      # 
      
      # this, because for curated Atlas, query, I don't need of the other meta data, 
      # I can add it later from the full meta data 
      select(sample_id, cell_id, file_id_cellNexus_single_cell, cell_type_unified_ensemble) |>
      filter(cell_type_unified_ensemble!="other", !cell_type_unified_ensemble |> is.na()) |> 
      filter(!file_id_cellNexus_single_cell |> is.na()) |> 
      
      # THIS SHOULD BE REMOVED IN THE FUTURE
      # mutate(file_id_cellNexus_single_cell = paste0(file_id_cellNexus_single_cell, ".h5ad")) |> 
      
      # NEST 
      as_tibble() |> 
      mutate(chunk = file_id_cellNexus_single_cell) |> 
      nest(data = -c(chunk)) |> 
      
      mutate(number_of_cells = map_int(data, nrow)) 
    
  }
  
  get_sce = 	function(tissue_cell_type_metadata, cache.path) {
    
    dir.create(cache.path, recursive = TRUE, showWarnings = FALSE)
    
    tissue_cell_type_metadata |>
      mutate(data = map2(
        data, chunk,
        ~ {
          
          file_id_cellNexus_single_cell = .y
          
          # THIS IS TEMPORARY BECAUSE I HAVE BRAIN DATASETS WITH 1M CELL THAT ARE HARD TO SAVE IN THE DB
          if(!file.exists(paste0(cache.path, "/X/", file_id_cellNexus_single_cell))) {
            warning(paste0(cache.path, "/X/", file_id_cellNexus_single_cell, " does not exist in the cache") )
            return(NULL)
          }
          
          .x |>
            CuratedAtlasQueryR:::get_data_container(
              repository = NULL, 
              cache_directory = cache.path, 
              grouping_column = "file_id_cellNexus_single_cell", 
              assays = "X"
            )
        }
        
        
      )) |> 
      
      # THIS IS TEMPORARY BECAUSE I HAVE BRAIN DATASETS WITH 1M CELL THAT ARE HARD TO SAVE IN THE DB
      filter(!map_lgl(data, is.null) ) |> 
      filter(map_int(data, ncol)>0)
    
  }
  
  get_pseudobulk = 	function(sce_df, external_directory) {
    
    # Parallelise
    cores = as.numeric(Sys.getenv("SLURM_CPUS_PER_TASK", unset = 1))
    bp <- MulticoreParam(workers = cores , progressbar = TRUE)  # Adjust the number of workers as needed
    
    dir.create(external_directory, recursive = TRUE, showWarnings = FALSE)
    
    sce_df |>
      
      
      mutate(data = map2(
        data, number_of_cells,
        ~ {
          
          assay_name = .x |>  assays() |> names() |> _[1]
          
          
          pseudobulk = 
            aggregateAcrossCells(
              .x, 
              colData(.x)[,c("sample_id", "cell_type_unified_ensemble")], 
              BPPARAM = bp, 
              use.assay.type = assay_name
            )
          colnames(pseudobulk) = paste0(colData(pseudobulk)$sample_id, "___", colData(pseudobulk)$cell_type_unified_ensemble)
          
          pseudobulk = pseudobulk |> select(.cell, sample_id, file_id_cellNexus_single_cell, cell_type_unified_ensemble, ncells)
          
          # Decrease size
          # We will reattach rowames later
          assay(pseudobulk) = assay(pseudobulk) |>  as("sparseMatrix")
          
          pseudobulk |> 
            HDF5Array::saveHDF5SummarizedExperiment(
              paste0(external_directory, "/", digest::digest(pseudobulk)),
              as.sparse = TRUE,
              replace = TRUE
            )
          
          
        }
      ))
    
  }
  
  aggregate = function(se_df, se_rownames){
    
    print("Start aggregate")
    gc()
    
    
    se_df |>
      
      # Add columns and filter
      mutate(data = map2(
        data, file_id_cellNexus_single_cell,
        ~ {
          
          # ## TEMPORARY FIX BECAUSE I FORGOT TO ADD ASSAY NAME
          # assays(.x) = .x |> assays() |> as.list() |>  set_names("counts")
          # #####
          
          # Add columns
          se = 
            .x |>
            mutate(file_id_cellNexus_single_cell = .y) |>
            select(-any_of(c("file_id_cellNexus_single_cell", ".cell", "original_cell_id")))
          
          
          # # Identify samples with many genes
          # sample_with_many_genes =
          #   se |>
          #   assay("counts") |>
          #   apply(2, function(x) (x>0) |> which() |> length()) |>
          #   enframe() |>
          #   mutate(name = as.character(name)) |>
          #   filter(value > 5000) |>
          #   pull(name)
          # se = se[,sample_with_many_genes, drop=FALSE]
          # 
          # # Filter samples with too few cells
          # se = se |> filter(.aggregated_cells > 10)
          
          se
          
        }, 
        .progress=TRUE
      )) |>
      
      nest(data = -file_id_cellNexus_single_cell) |> 
      mutate(data = map(data,
                        ~ {
                          se = .x |> 
                            pull(data) |> 
                            do.call(cbind, args= _)
                          
                          # # Filter very rare gene-transcripts
                          # all_zeros = assay(se, "counts") |> rowSums() |> equals(0) 
                          # se = se[!all_zeros,]
                          # lower_than_total_counts = assay(se, "counts") |> rowSums() < 15 
                          # se = se[!lower_than_total_counts,]
                          
                          
                          # Make it sparse
                          se@assays@data$counts = se@assays@data$counts |> as("sparseMatrix")
                          
                          # Attach rownames back
                          #rownames(se) = se_rownames
                          
                          se
                        })
      )
    
  }
  
  #' Process a SummarizedExperiment Object
  #'
  #' This function processes a single \code{SummarizedExperiment} object by adjusting its assay to include all genes,
  #' filling missing genes with zeros, and recording gene presence per sample as a separate assay.
  #' It returns a \code{SummarizedExperiment} with the adjusted assay and gene presence information as a secondary assay.
  #'
  #' **Why This Function Is Needed:**
  #'
  #' When combining multiple \code{SummarizedExperiment} objects, each may have different gene sets.
  #' This function standardizes each object's assay to include all common genes, facilitating seamless combination.
  #' By returning a \code{SummarizedExperiment}, it simplifies downstream processing and integration.
  #'
  #' @param se A \code{SummarizedExperiment} object to be processed.
  #' @param gene_index A named integer vector mapping gene names to their indices.
  #' @param se_id A unique identifier for the \code{SummarizedExperiment}, used to ensure unique sample names.
  #' @param external_directory A directory path to save the processed \code{SummarizedExperiment}.
  #' @return A \code{SummarizedExperiment} object with:
  #'   \item{assays}{A list containing the adjusted assay with all genes, named using the original assay name, and a "gene_presence" assay.}
  #'   \item{colData}{The column data from the original \code{SummarizedExperiment}, with updated sample names.}
  #'
  #' @examples
  #' # Process a single SummarizedExperiment
  #' processed_se <- process_se(se, gene_index, se_id = 1, external_directory = "output_directory")
  #'
  #' @importFrom SummarizedExperiment assay
  #' @importFrom SummarizedExperiment assayNames
  #' @importFrom SummarizedExperiment colData
  #' @importFrom SummarizedExperiment SummarizedExperiment
  #' @importFrom S4Vectors DataFrame
  #' @importFrom Matrix sparseMatrix
  #' @importFrom HDF5Array saveHDF5SummarizedExperiment
  #' @importFrom stringr str_remove
  #' @export
  process_se <- function(se, gene_index, se_id, external_directory) {
    
    dir.create(external_directory, recursive = TRUE, showWarnings = FALSE)
    
    all_genes <- names(gene_index)
    n_genes <- length(gene_index)
    genes_se <- rownames(se)
    n_cells <- ncol(se)
    
    # Update sample names to be unique by prefixing with se_id
    sample_names = colnames(se)
    
    # Get indices of genes in the full gene list
    idx <- gene_index[genes_se]
    
    # Get the assay data and convert to sparse TsparseMatrix if not already
    assay_se <- assay(se)
    if (!is(assay_se, "TsparseMatrix")) {
      assay_se <- as(as(assay_se, "sparseMatrix"), "TsparseMatrix")
    }
    
    # Adjust indices
    i_indices <- idx[assay_se@i + 1]
    j_indices <- assay_se@j + 1
    
    # Create sparse matrix for the assay with all genes
    assay_full <- sparseMatrix(
      i = i_indices,
      j = j_indices,
      x = assay_se@x,
      dims = c(n_genes, n_cells),
      dimnames = list(all_genes, sample_names)
    )
    
    # Record presence of genes per sample in a sparse binary matrix
    gene_presence_i <- Matrix(FALSE, nrow = n_genes, ncol = n_cells, sparse = TRUE)
    gene_presence_i[idx, ] <- TRUE
    colnames(gene_presence_i) <- sample_names
    rownames(gene_presence_i) <- all_genes
    
    # Get the original assay name(s)
    assay_name <- assayNames(se)[1]  # Assuming there is only one assay
    
    # Create processed SummarizedExperiment with original assay name and "gene_presence"
    processed_se <- SummarizedExperiment(
      assays = setNames(list(assay_full, gene_presence_i), c(assay_name, "gene_presence")),
      colData = colData(se)
    )
    
    # Save the processed SummarizedExperiment to an external file
    se = HDF5Array::saveHDF5SummarizedExperiment(
      processed_se,
      paste0(external_directory, "/", str_remove(se_id, "\\.h5ad")),
      as.sparse = TRUE,
      replace = TRUE
    )
    
    se
  }
  
  #-----------------------#
  # Pipeline
  #-----------------------#
  list(
    tar_target(cache_directory, "/vast/scratch/users/mangiola.s/cellNexus/cellxgene/26_11_2024", deployment = "main"),
    
    tar_target(
      metadata_parquet, 
      "/vast/projects/cellxgene_curated/cellNexus/cell_metadata_cell_type_consensus_v1_0_4.parquet", 
      format = "file", 
      deployment = "main"
    ),
    
    # Do metadata
    tar_target(
      metadata_split,
      split_metadata(metadata_parquet),
      resources = tar_resources(crew = tar_resources_crew("slurm_1_20")), 
      packages = c("arrow", "dplyr",  "duckdb", "tidyr", "dplyr", "purrr", "glue")
    ),
    
    # Get SCE SMALL
    tar_target(
      metadata_split_pseudobulk,
      metadata_split |>  get_sce(paste(cache_directory, "/single_cell/")) |> get_pseudobulk(paste0(result_directory, "/external/")),
      pattern = map(metadata_split),
      resources = tar_resources(crew = tar_resources_crew("elastic"))
    ) ,
    
    tar_target(
      all_genes,
      {
        all_genes = metadata_split_pseudobulk |> pull(data) |> map(rownames) |> unlist() |> unique() 
        all_genes = setNames(seq_len(length(all_genes)), all_genes)
        all_genes
      },
      #iteration = "list",
      resources = tar_resources(crew = tar_resources_crew("slurm_1_150"))
      
    ),

    tar_target(
      metadata_grouped_pseudobulk_processed,
      {
        if(nrow(metadata_split_pseudobulk) == 0) return(NULL)
        
        process_se(
          metadata_split_pseudobulk |> pull(data) |> _[[1]], 
          all_genes, 
          metadata_split_pseudobulk |> pull(chunk) |> _[[1]], 
          paste0(cache_directory, "/pseudobulk/counts/")
        )
      },
      pattern = map(metadata_split_pseudobulk),
      iteration = "list",
      resources = tar_resources(crew = tar_resources_crew("elastic"))
    ),
    
    # SPLIT 1
    tar_target(
      metadata_grouped_pseudobulk_processed_split_1,
      metadata_grouped_pseudobulk_processed |> 
        split(ceiling(seq_along(metadata_grouped_pseudobulk_processed) / 10)) ,
      iteration = "list",
      resources = tar_resources(crew = tar_resources_crew("slurm_1_80"))
    ),
    tar_target(
      metadata_grouped_pseudobulk_processed_split_1_HDF5,
      {
        
        se = 
          metadata_grouped_pseudobulk_processed_split_1 |> 
          
          # !!!
          # THIS SHOULD NOT BE NEEDED AND SHOULD BE RESOLVED IN TRANFORM DATA
          # !!!
          purrr::discard(is.null) |>
          
          # !!!
          # THIS SHOULD NOT BE NEEDED AND SHOULD BE RESOLVED IN TRANFORM DATA
          # !!!
          map(~ { names(assays(.x))[1] = "X"; .x}) |> 
          

          # Bind
          do.call(cbind, args = _) 
        
        if(length(se) == 0) return(NULL)
        
        se |> 
          HDF5Array::saveHDF5SummarizedExperiment(
            paste0(result_directory, "/external", digest::digest(metadata_grouped_pseudobulk_processed_split_1)),
            as.sparse = TRUE,
            replace = TRUE,
            verbose = TRUE
          )
      }
       ,
      pattern = map(metadata_grouped_pseudobulk_processed_split_1),
      iteration = "list",
      resources = tar_resources(crew = tar_resources_crew("elastic"))
    ),
    
    # SPLIT 2
    tar_target(
      metadata_grouped_pseudobulk_processed_split_2,
      metadata_grouped_pseudobulk_processed_split_1_HDF5 |> 
        split(ceiling(seq_along(metadata_grouped_pseudobulk_processed_split_1) / 10)) ,
      iteration = "list",
      resources = tar_resources(crew = tar_resources_crew("slurm_1_80"))
    ),
    tar_target(
      metadata_grouped_pseudobulk_processed_split_2_HDF5,
      {
        se = 
          metadata_grouped_pseudobulk_processed_split_2 |> 
          
          # !!!
          # THIS SHOULD NOT BE NEEDED AND SHOULD BE RESOLVED IN TRANFORM DATA
          # !!!
          purrr::discard(is.null) |>
          
          # !!!
          # THIS SHOULD NOT BE NEEDED AND SHOULD BE RESOLVED IN TRANFORM DATA
          # !!!
          map(~ { names(assays(.x))[1] = "X"; .x}) |> 
          
          
          
          # Bind
          do.call(cbind, args = _) 
        
        if(length(se) == 0) return(NULL)
        
        se |>
          HDF5Array::saveHDF5SummarizedExperiment(
            paste0(result_directory, "/external", digest::digest(se)),
            as.sparse = TRUE,
            replace = TRUE,
            verbose = TRUE, 
          )
      }
       ,
      pattern = map(metadata_grouped_pseudobulk_processed_split_2),
      iteration = "list",
      resources = tar_resources(crew = tar_resources_crew("elastic"))
    ),
    
    # SPLIT 3
    tar_target(
      metadata_grouped_pseudobulk_processed_split_3,
      metadata_grouped_pseudobulk_processed_split_2_HDF5 |> 
        split(ceiling(seq_along(metadata_grouped_pseudobulk_processed_split_2) / 10)) ,
      iteration = "list",
      resources = tar_resources(crew = tar_resources_crew("slurm_1_80"))
    ),
    tar_target(
      metadata_grouped_pseudobulk_processed_split_3_HDF5,
      {
        se = 
          metadata_grouped_pseudobulk_processed_split_3 |> 
          
          # !!!
          # THIS SHOULD NOT BE NEEDED AND SHOULD BE RESOLVED IN TRANFORM DATA
          # !!!
          purrr::discard(is.null) |>
          
          # !!!
          # THIS SHOULD NOT BE NEEDED AND SHOULD BE RESOLVED IN TRANFORM DATA
          # !!!
          map(~ { names(assays(.x))[1] = "X"; .x}) |> 
          
          # Bind
          do.call(cbind, args = _) 
        
        if(length(se) == 0) return(NULL)
        
        se |>
          HDF5Array::saveHDF5SummarizedExperiment(
            paste0(result_directory, "/external", digest::digest(se)),
            as.sparse = TRUE,
            replace = TRUE,
            verbose = TRUE, 
          )
      } ,
      pattern = map(metadata_grouped_pseudobulk_processed_split_3),
      iteration = "list",
      resources = tar_resources(crew = tar_resources_crew("elastic"))
    )
    
  )
  
  
}, 
ask = FALSE, 
script = glue("{result_directory}/_targets.R")
)

job::job({
  
  tar_make(
   # callr_function = NULL,
    reporter = "summary",
    script = glue("{result_directory}/_targets.R"),
    store = glue("{result_directory}/_targets")
  )
  
})

x = tar_read(metadata_grouped_pseudobulk_processed_split_3_HDF5, store = glue("{result_directory}/_targets"))
x = tar_read(metadata_split, store = glue("{result_directory}/_targets"), branches = 1)


tar_meta(store = glue("{result_directory}/_targets")) |> 
  arrange(desc(time)) |>
  filter(!error |> is.na()) |> 
  select(name, error)

tar_workspace(
  "metadata_grouped_pseudobulk_processed_split_2_HDF5_c6dd2f9db8a49dd1", 
  script = glue("{result_directory}/_targets.R"),
  store = glue("{result_directory}/_targets")
)

x = tar_read(metadata_grouped_pseudobulk_processed_split_1_HDF5, store = glue("{result_directory}/_targets"))

c(
  # "metadata_split",
  # "metadata_split_SMALL",
  # "metadata_split_BIG"
  # "metadata_split_pseudobulk_SMALL",
  # "metadata_split_pseudobulk_BIG",
  # "metadata_grouped_pseudobulk",
  # "metadata_grouped_pseudobulk_processed",
  # "metadata_grouped_pseudobulk_processed_split_1",
  "metadata_grouped_pseudobulk_processed_split_1_HDF5",
  "metadata_grouped_pseudobulk_processed_split_2",
  "metadata_grouped_pseudobulk_processed_split_2_HDF5",
  "metadata_grouped_pseudobulk_processed_split_3",
  "metadata_grouped_pseudobulk_processed_split_3_HDF5"
) |> 
  tar_invalidate(glue("{result_directory}/_targets"))

tar_meta(store = glue("{result_directory}/_targets")) |> 
  filter(!error |> is.na()) |> 
  arrange(desc(time)) |> 
  select(name, error) |>  
  View()



# # Save into one SE
#   library(crew.cluster)
#   controller <- crew_controller_slurm(
#     name = "elastic",
#     seconds_idle = 30,
#     options_cluster = crew_options_slurm(
#       memory_gigabytes_required = c(160), 
#       cpus_per_task = c(40), 
#       time_minutes = c(60*24),
#       verbose = T
#     )
#   )
#   controller$start()
#   controller$push(name = "aggregate", command = {
    
    library(SummarizedExperiment)
    library(tidySummarizedExperiment)
    
    result_directory = "/vast/scratch/users/mangiola.s/pseudobulk_cellNexus_1_0_4"
    
    pseudobulk = 
      targets::tar_read(metadata_grouped_pseudobulk_processed_split_3_HDF5, store = glue::glue("{result_directory}/_targets"))  |>
      do.call(cbind, args = _) 
    
    
    duplicated_samples = 
      pseudobulk |> colnames() |>
      _[pseudobulk |> colnames() |> duplicated()] |> 
      unique()
    
    duplicated_samples_split <- split(
      duplicated_samples,
      cut(seq_along(duplicated_samples), breaks = 5, labels = FALSE)
    )
    

    
# This aggregate duplicate samples due to different split between cell type unified and consensus 
job::job({
  
  i = 5
  
  pseudobulk_duplicated = pseudobulk[, colnames(pseudobulk) %in% duplicated_samples_split[[i]] ]
  
  
  # Parallelise
  cores = as.numeric(Sys.getenv("SLURM_CPUS_PER_TASK", unset = 1)) - 2
  bp <- BiocParallel::MulticoreParam(workers = cores , progressbar = TRUE)  # Adjust the number of workers as needed
  
  se_duplicated = 
    scuttle::aggregateAcrossCells(
      pseudobulk_duplicated, 
      colData(pseudobulk_duplicated)[,c("sample_id", "cell_type_unified_ensemble")], 
      BPPARAM = bp, 
      use.assay.type = c("X", "gene_presence")
    )
  
  
  se_duplicated |> assays() = se_duplicated |> assays() |> purrr::map(Matrix, sparse = TRUE) 
  colData(se_duplicated) = colData(se_duplicated)[,-duplicated(colnames(colData(se_duplicated)))]
  colData(se_duplicated) = colData(se_duplicated)[,! colnames(colData(se_duplicated)) %in% "ncells"]
  colData(se_duplicated) = colData(se_duplicated)[,colnames(colData(pseudobulk))]
  
  se_duplicated |> 
    HDF5Array::saveHDF5SummarizedExperiment(
    glue::glue("/vast/scratch/users/mangiola.s/pseudobulk_cellNexus_1_0_4/deduplicated_pseudobulk___{i}"),
    as.sparse = TRUE,
    replace = TRUE,
    verbose = TRUE
  )
})
    
pseudobulk_deduplicated = 
  glue::glue("/vast/scratch/users/mangiola.s/pseudobulk_cellNexus_1_0_4/deduplicated_pseudobulk___{1:5}") |> 
  lapply(HDF5Array::loadHDF5SummarizedExperiment) |> 
  do.call(cbind, args = _) |> 
  cbind( pseudobulk[, !colnames(pseudobulk) %in% unlist(duplicated_samples_split)])

colnames(pseudobulk_deduplicated) = paste0(colData(pseudobulk_deduplicated)$sample_id, "___", colData(pseudobulk_deduplicated)$cell_type_unified_ensemble)

names(assays(pseudobulk_deduplicated))[1] = "counts"

pseudobulk_deduplicated = 
  pseudobulk_deduplicated |> 
  
  left_join(
    readRDS("/vast/projects/mangiola_immune_map/PostDoc/immuneHealthyBodyMap/sccomp_on_cellNexus_1_0_1_regularised/cell_metadata_1_0_1_sccomp_input_counts.rds") |> 
      mutate(do_analyse = TRUE)
  ) 

# Attach info
pseudobulk_deduplicated |> colData() = pseudobulk_deduplicated |> colData() |> as_tibble(rownames = "rn") |> mutate(do_analyse = if_else(do_analyse |> is.na(), FALSE, do_analyse)) |> data.frame(row.names = "rn") |> DataFrame()
pseudobulk_deduplicated |> assay("gene_presence") = pseudobulk_deduplicated |> assay("gene_presence") > 0
system("~/bin/rclone copy box_adelaide:/Mangiola_ImmuneAtlas/taskforce_shared_folder/selected_gene_list.csv /vast/projects/mangiola_immune_map/PostDoc/immuneHealthyBodyMap/ ")
rowData(pseudobulk_deduplicated)$is_gene_shared = rownames(pseudobulk_deduplicated) %in% readr::read_csv("/vast/projects/mangiola_immune_map/PostDoc/immuneHealthyBodyMap/selected_gene_list.csv")$Selected_genes

job::job({
  setAutoBlockSize(size=1e9)
  
  pseudobulk_deduplicated = 
    pseudobulk_deduplicated |> 
    
    HDF5Array::saveHDF5SummarizedExperiment(
      "/vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample_cell_type",
      as.sparse = TRUE,
      replace = TRUE,
      verbose = TRUE
    )
})

system("~/bin/rclone copy /vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample_cell_type box_adelaide:/Mangiola_ImmuneAtlas/taskforce_shared_folder/")

# Save as anndata


job::job({
  zellkonverter::writeH5AD(
    se, 
    verbose = TRUE, 
    file = "/vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample_cell_type_1_0_4.h5ad", 
    X_name = "counts"
  )
})

#assays(se) = assays(se)[1] # aggregateAcrossCells does not like booleans

library(duckdb)
library(duckdb)
is_immune_tbl =
  tbl(
  dbConnect(duckdb::duckdb(), dbdir = ":memory:"),
  sql("SELECT * FROM read_parquet('/vast/projects/cellxgene_curated/cellNexus/cell_metadata_cell_type_consensus_v1_0_4.parquet')")
) |> 
  distinct(cell_type_unified_ensemble, is_immune) |> 
  as_tibble()

pseudobulk_deduplicated |> colData() = pseudobulk_deduplicated |> colData() |> as_tibble(rownames = "rn") |> left_join(is_immune_tbl) |> data.frame(row.names = "rn") |> DataFrame()
pseudobulk_deduplicated |> colData() = pseudobulk_deduplicated |> colData() |> as_tibble(rownames = "rn") |> mutate(is_immune = is_immune |> as.integer()) |> data.frame(row.names = "rn") |> DataFrame()

job::job({
  
  library(SummarizedExperiment)
  library(tidySummarizedExperiment)
  
  # Parallelise
  cores = as.numeric(Sys.getenv("SLURM_CPUS_PER_TASK", unset = 1)) - 2
  cores = 20
  bp <- BiocParallel::MulticoreParam(workers = cores , progressbar = TRUE)  # Adjust the number of workers as needed
  
  
  pseudobulk_sample = 
    pseudobulk_deduplicated |> 
    scuttle::aggregateAcrossCells(
      colData(pseudobulk_deduplicated)[,c("sample_id", "is_immune")], 
      use.assay.type = c("counts", "gene_presence"),  # Specify the correct assay name if there are multiple assays
      BPPARAM = bp
    )

  colnames(pseudobulk_sample) = paste0(colData(pseudobulk_sample)$sample_id, "___", colData(pseudobulk_sample)$is_immune)
  pseudobulk_sample |> colData() = pseudobulk_sample |> colData() |> as_tibble(rownames = "rn") |> mutate(do_analyse = if_else(do_analyse |> is.na(), TRUE, do_analyse)) |> data.frame(row.names = "rn") |> DataFrame()
  
  pseudobulk_sample |> colData() = 
    pseudobulk_sample |> colData() |> _[, c("sample_id", "is_immune", "do_analyse")] |> as_tibble(rownames = "rn") |> 
    left_join(
      readRDS("/vast/projects/mangiola_immune_map/PostDoc/immuneHealthyBodyMap/sccomp_on_cellNexus_1_0_1_regularised/cell_metadata_1_0_1_sccomp_input_counts.rds") |> 
        distinct(sample_id, donor_id, title, dataset_id, collection_id, age_days, sex, ethnicity_groups, tissue_groups, assay_groups, disease_groups, age_bin_sex_specific) 
    ) |> 
    data.frame(row.names = "rn") |> 
    DataFrame() 
  
  setAutoBlockSize(size=1e9)
  
  pseudobulk_sample = 
    pseudobulk_sample |> 
    HDF5Array::saveHDF5SummarizedExperiment(
      "/vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample_is_immune",
      as.sparse = TRUE,
      replace = TRUE,
      verbose = TRUE
    )
})


pseudobulk_sample = HDF5Array::loadHDF5SummarizedExperiment("/vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample_is_immune")
system("~/bin/rclone copy /vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample_is_immune/ box_adelaide:/Mangiola_ImmuneAtlas/taskforce_shared_folder/pseudobulk_sample_is_immune/")

pseudobulk_sample |> zellkonverter::writeH5AD(
  "/vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample_is_immune.h5ad", 
  verbose = TRUE, 
  X_name = "counts",
  compression = "gzip"
)

# THIS IS BECAUSE HDF5 IS NOT PORTABLE
pseudobulk_sample |> assays() = pseudobulk_sample |> assays() |> map(~{

  # Convert to sparse matrix while preserving rownames and colnames
  sparse_matrix <- as(.x, "sparseMatrix")  
  
  # # Preserve rownames and colnames
  # rownames(sparse_matrix) <- rownames(.x)
  # colnames(sparse_matrix) <- colnames(.x)
  
  sparse_matrix
  
}, .progress  = TRUE)
  

pseudobulk_sample |> saveRDS("/vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample_is_immune.rds")
system("~/bin/rclone copy /vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample_is_immune.rds box_adelaide:/Mangiola_ImmuneAtlas/taskforce_shared_folder/")



system("~/bin/rclone copy /vast/projects/cellxgene_curated/cellNexus/pseudobulk_sample/ box_adelaide:/Mangiola_ImmuneAtlas/taskforce_shared_folder/pseudobulk_sample/")


tar_meta(starts_with("metadata_split_SMALL"), store = glue("{result_directory}/_targets"))

# tar_invalidate(metadata_split, store = glue("{result_directory}/_targets"))
# tar_invalidate(metadata_grouped_pseudobulk_processed_split_1, store = glue("{result_directory}/_targets"))
